fix a layering issue; add proxy stub for variables
