+++ src/vs/workbench/api/common/positron/extHostLanguageRuntime.ts
@@ -21,8 +21,8 @@ import { SerializableObjectWithBuffers } from '../../../services/extensions/comm
 import { VSBuffer } from '../../../../base/common/buffer.js';
 import { generateUuid } from '../../../../base/common/uuid.js';
 import { CancellationToken } from '../../../../base/common/cancellation.js';
-import { ILanguageRuntimeCodeExecutedEvent } from '../../../services/positronConsole/browser/interfaces/positronConsoleService.js';
 import { Variable } from '../../../services/languageRuntime/common/positronVariablesComm.js';
+import { ILanguageRuntimeCodeExecutedEvent } from '../../../services/positronConsole/common/positronConsoleCodeExecution.js';
 
 /**
  * Interface for code execution observers
@@ -1244,7 +1244,12 @@ export class ExtHostLanguageRuntime implements extHostProtocol.ExtHostLanguageRu
 
 	public getSessionVariables(sessionId: string, accessKeys?: Array<Array<string>>):
 		Promise<Variable[]> {
-		return Promise.resolve([]);
+		for (let i = 0; i < this._runtimeSessions.length; i++) {
+			if (this._runtimeSessions[i].metadata.sessionId === sessionId) {
+				return this._proxy.$getSessionVariables(i, accessKeys);
+			}
+		}
+		throw new Error(`Session with ID '${sessionId}' not found`);
 	}
 
 	/**
